I am now focused on making a fishing simulator.Today I created the fish and fishGenerator objects and a main class to run the code. I renamed the runner class to the name of the project plus runner. I imported methods like lists and arrays and began development on the time aspect of my project.

+++ src/Fish.java
@@ -1,9 +1,9 @@
 public class Fish {
     private final String name;
-    private final int weight;
-    private final int length;
+    private final double weight;
+    private final double length;
 
-    public Fish (String name , int weight, int length){
+    public Fish (String name , double weight, double length){
         this.name = name;
         this.weight = weight;
         this.length = length;
@@ -11,10 +11,10 @@ public class Fish {
     public String getName(){
       return name;
     }
-    public int getWeight(){
+    public double getWeight(){
         return weight;
     }
-    public int getLength(){
+    public double getLength(){
         return length;
     }
 }
